libraries/bmh_jmetalpy_extensions/pyproject.toml
[project]
name = "bmh_jmetalpy_extensions"
version = "0.0.1"
description = "jMetalPy Extensions"
readme = "README.md"
authors = [
    { name = "Michael Cipold", email = "github@cipold.de" }
]
requires-python = ">=3.10"
dependencies = [
    "dask",
    "jmetalpy==1.5.5",
]
license = { text = "MIT License" }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
